Update remediation content
This makes some of the relevant parts of the code examples appear in bold, so
that it's easier to see what the differences are. We found that it was easy to
miss the part of the code that the example was trying to show.

It also updates some of the text to make the advice clearer, and fixes some
typos.